Move data gen mods to Forge extension
