+++ src/main/java/net/fabricmc/loom/extension/ForgeExtensionImpl.java
@@ -24,24 +24,35 @@
 
 package net.fabricmc.loom.extension;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
 import javax.inject.Inject;
 
+import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.SetProperty;
 
+import net.fabricmc.loom.LoomGradleExtension;
 import net.fabricmc.loom.api.ForgeExtensionAPI;
+import net.fabricmc.loom.configuration.ide.RunConfigSettings;
 
 public class ForgeExtensionImpl implements ForgeExtensionAPI {
+	private final LoomGradleExtension extension;
 	private final Property<Boolean> convertAccessWideners;
 	private final SetProperty<String> extraAccessWideners;
 	private final ConfigurableFileCollection accessTransformers;
 	private final SetProperty<String> mixinConfigs;
 	private final Property<Boolean> useFabricMixin;
+	private final List<String> dataGenMods = new ArrayList<>(); // not a property because it has custom adding logic
 
 	@Inject
-	public ForgeExtensionImpl(Project project) {
+	public ForgeExtensionImpl(Project project, LoomGradleExtension extension) {
+		this.extension = extension;
 		convertAccessWideners = project.getObjects().property(Boolean.class).convention(false);
 		extraAccessWideners = project.getObjects().setProperty(String.class).empty();
 		accessTransformers = project.getObjects().fileCollection();
@@ -82,5 +93,25 @@ public class ForgeExtensionImpl implements ForgeExtensionAPI {
 	@Override
 	public Property<Boolean> getUseFabricMixin() {
 		return useFabricMixin;
+	}
+
+	@Override
+	public List<String> getDataGenMods() {
+		return Collections.unmodifiableList(dataGenMods);
+	}
+
+	@SuppressWarnings("Convert2Lambda")
+	@Override
+	public void dataGen(Action<DataGenConsumer> action) {
+		action.execute(new DataGenConsumer() {
+			@Override
+			public void mod(String... modIds) {
+				dataGenMods.addAll(Arrays.asList(modIds));
+
+				if (modIds.length > 0 && extension.getRunConfigs().findByName("data") == null) {
+					extension.getRunConfigs().create("data", RunConfigSettings::data);
+				}
+			}
+		});
 	}
 }
